backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
email-validator
bcrypt
pyjwt
python-multipart
alembic
Pillow
python-magic
pytest
pytest-asyncio
httpx
